pyproject.toml
[tool.poetry]
name = "bechdelai"
version = "0.0.2"
description = "Automating the Bechdel test and its variants for feminine representation in movies with AI"
authors = ["Data For Good <hellodataforgood@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<3.9.7 || >3.9.7,<3.11"
jupyter = "^1.0.0"
pandas = "^1.3.4"
numpy = "^1.21.3"
scikit-learn = "^1.0.1"
matplotlib = "^3.4.3"
plotly = "^5.3.1"
opencv-python = "^4.5.4"
tqdm = "^4.62.3"
tensorflow = "^2.7.0"
deepface = "^0.0.68"
umap-learn = "^0.5.2"
retina-face = "^0.0.12"
transformers = "^4.17.0"
torch = "^1.11.0"
torchvision = "^0.12.0"
torchaudio = "^0.11.0"
ftfy = "^6.1.1"
python-dotenv = ">=0.19.2"
lxml = "^4.8.0"
PyMuPDF = "^1.19.6"
Click = ">=8.0.4"
requests = ">=2.26.0"
beautifulsoup4 = ">=4.10.0"
chardet = ">=4.0.0"
openpyxl = "^3.0.10"
pytube = "^12.1.0"
mediapipe = "^0.9.0"
scenedetect = {extras = ["opencv"], version = "^0.6.1"}
moviepy = "^1.0.3"
SpeechRecognition = "^3.8.1"
inaSpeechSegmenter = "^0.7.3"
llvmlite = "^0.39.1"
setuptools = ">=62.3.2"
openai = "^0.27.0"
openai-whisper = "^20230307"
gradio = "^3.20.1"

[tool.poetry.dev-dependencies]
pywin32 = { version = "305", markers = "sys_platform == 'win32'" }
jupyter = "^1.0.0"
mkdocs = "^1.4.1"
streamlit = "1.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
scrap-movies = 'bechdelai.scripts.scrap_movies:cli'
